Specify tx pointer recovery
The tx pointer records the current state of transaction execution or in
other words the next transaction the system is supposed to execute.
Ideally, this value would be tracked by the chain itself, but this would
require a chain fork. Instead, we have to rely on consensus among the
keypers.

In practice, under normal conditions, the keypers use the txPointer in
`DecryptionKeys` messages as canonical (note that in order to be valid,
these messages need to be signed off by at least a threshold of the
keypers). However, if no such message is received for a long time, key
generation must be failing. This might be due to keypers not agreeing on
the txPointer. In order to recover from this failure, the keypers reset
the txPointer to the number of encrypted transactions in the queue,
i.e., they jump to the end (because this is the easiest value to agree
on without communication).

Keypers should do this when key generation has been failing for some
time. What "some time" means is left to the keypers to decide. It is a
tradeoff between quick recovery and poor UX due to skipped transactions.

Note that this automatically contains a retry mechanism: If key
production is failing, no decryption keys messages will be received, so
they also will not update the tx pointer.